update and fixes

+++ vwr.sh
@@ -10,11 +10,9 @@ re='^[0-9]+$'
 v="$1"
 now="$2" # es el nombre de la lista (file name)
 nuw="$3" # es el numero de posicion del item (file name) en la lista
-if [ "$v" = v1 ]; then
-	ind="$DC_tlt/cnfg1"
-elif [ "$v" = v2 ]; then
-	ind="$DC_tlt/cnfg2"
-fi
+[[ "$v" = v1 ]] && ind="$DC_tlt/cnfg1"
+[[ "$v" = v2 ]] && ind="$DC_tlt/cnfg2"
+
 if ! [[ $nuw =~ $re ]]; then
 	nuw=$(cat "$ind" \
 	| grep -Fxon "$now" \
@@ -28,58 +26,45 @@ if [ -z "$nme" ]; then
 fi
 
 if [ -f "$DM_tlt/words/$nme.mp3" ]; then
-	file="$DM_tlt/words/$nme.mp3"
-	tgs=$(eyeD3 "$file")
+	tgs=$(eyeD3 "$DM_tlt/words/$nme.mp3")
 	trgt="$nme"
 	src=$(echo "$tgs" | grep -o -P '(?<=IWI2I0I).*(?=IWI2I0I)')
 	exmp=$(echo "$tgs" | grep -o -P '(?<=IWI3I0I).*(?=IWI3I0I)' | tr '_' '\n')
 	mrk=$(echo "$tgs" | grep -o -P '(?<=IWI4I0I).*(?=IWI4I0I)')
-	exm1=$(echo "$exmp" | sed -n 1p)
-	dftn=$(echo "$exmp" | sed -n 2p)
-	ntes=$(echo "$exmp" | sed -n 3p)
-	dfnts="--field=<span color='#696464'>$dftn</span>\\n:lbl"
-	ntess="--field=<span color='#868686'>$ntes</span>\\n:lbl"
+	[[ $(echo "$exmp" | sed -n 2p) ]] \
+	&& dfnts="--field=<span color='#696464'>$(echo "$exmp" | sed -n 2p)</span>\\n:lbl"
+	[[ $(echo "$exmp" | sed -n 3p) ]] \
+	&& ntess="--field=<span color='#868686'>$(echo "$exmp" | sed -n 3p)</span>\\n:lbl"
 	hlgt=$(echo $trgt | awk '{print tolower($0)}')
-	exmp1=$(echo "$exm1" \
-	| sed "s/"$hlgt"/<span background='#CFFF8B'>"$hlgt"<\/\span>/g")
-	if [ "$mrk" = TRUE ]; then
-		trgt=$(echo "<span color='#00335E'><b>"$trgt"</b></span>")
-	fi
-	if [ "$ap" = TRUE ]; then
-		(killall play & sleep 0.3 && play "$DM_tlt/words/$nme.mp3") &
-	fi
-
-	$yad --columns=1 --form \
-	--window-icon=idiomind --scroll --text-align=center \
+	exmp1=$(echo "$(echo "$exmp" | sed -n 1p)" | sed "s/"$(echo $trgt \
+	| awk '{print tolower($0)}')"/<span background='#CFFF8B'>"$(echo $trgt \
+	| awk '{print tolower($0)}')"<\/\span>/g")
+	[[ "$(echo "$tgs" | grep -o -P '(?<=IWI4I0I).*(?=IWI4I0I)')" = TRUE ]] \
+	&& trgt=$(echo "<big><u><b>"$trgt"</b></u></big>")
+	[[ "$ap" = TRUE ]] && (killall play & sleep 0.3 && play "$DM_tlt/words/$nme.mp3") &
+	
+	yad --form --window-icon=idiomind --scroll --text-align=center \
 	--skip-taskbar --center --title=" " --borders=10 \
 	--quoted-output --on-top --selectable-labels \
 	--text="<big><big><big>$trgt</big></big></big>\\n\\n<i>$src</i>\\n\\n" \
 	--field="":lbl \
-	--field="<i><span color='#808080'>$exmp1</span></i>\\n:lbl" "$dfnts" "$ntess" \
+	--field="<i><span color='#808080'>$exmp1</span></i>:lbl" "$dfnts" "$ntess" \
 	--width="$wth" --height="$eht" --center \
 	--button=gtk-edit:4 --button="Play":"play '$DM_tlt/words/$nme.mp3'" \
 	--button=gtk-go-up:3 --button=gtk-go-down:2 \
 	--dclick-action=/usr/share/idiomind/audio/pl >/dev/null 2>&1
 	
 elif [ -f "$DM_tlt/$nme.mp3" ]; then
-	file="$DM_tlt/$nme.mp3"
-	tgs=$(eyeD3 "$file")
-	if [ $(sed -n 4p $DC_s/cnfg1) = TRUE ]; then
-		trgt=$(echo "$tgs" | grep -o -P '(?<=IGMI3I0I).*(?=IGMI3I0I)')
-	else
-		trgt=$(echo "$tgs" | grep -o -P '(?<=ISI1I0I).*(?=ISI1I0I)')
-	fi
+	tgs=$(eyeD3 "$DM_tlt/$nme.mp3")
+	[[ $(sed -n 4p $DC_s/cnfg1) = TRUE ]] \
+	&& trgt=$(echo "$tgs" | grep -o -P '(?<=IGMI3I0I).*(?=IGMI3I0I)') \
+	|| trgt=$(echo "$tgs" | grep -o -P '(?<=ISI1I0I).*(?=ISI1I0I)')
 	src=$(echo "$tgs" | grep -o -P '(?<=ISI2I0I).*(?=ISI2I0I)')
 	lwrd=$(echo "$tgs" | grep -o -P '(?<=IPWI3I0I).*(?=IPWI3I0I)' | tr '_' '\n')
+	[[ ! -f "$DM_tlt/$nme.mp3" ]] && exit 1
+	[[ "$ap" = TRUE ]] && (killall play & sleep 0.3 && play "$DM_tlt/$nme.mp3") &
 	
-	if [ ! -f "$file" ]; then
-		exit 1
-	fi
-	if [ "$ap" = TRUE ]; then
-		(killall play & sleep 0.3 && play "$DM_tlt/$nme.mp3") &
-	fi
-	
-	echo "$lwrd" | $yad --list --print-column=0 --no-headers \
+	echo "$lwrd" | yad --list --print-column=0 --no-headers \
 	--window-icon=idiomind --scroll --text-align=center \
 	--skip-taskbar --center --title=" " --borders=10 \
 	--on-top --selectable-labels --expand-column=0 \
